order/receiptbackend

+++ Wood_and_Funiture_Management_System_Backend/src/main/java/com/group_project/wfms_backend/dto/auth/CustomerOrderRequestDTO.java
@@ -12,7 +12,7 @@ public class CustomerOrderRequestDTO {
     private BigDecimal paidAmount;
     private String status;           // Optional override
     private LocalDate orderDate;
-    private Long createdById;
+    private Integer createdById;
     private List<OrderDetailDTO> orderDetails;
 
     @Data
